requirements.txt
colorama==0.4.6
requests==2.31.0
dnspython==2.4.2
python-whois==0.8.0
beautifulsoup4==4.12.2
